Make this plugin hosted in our own Artifactory
The plugin is from an officiel Github site but it was published
at a privately hosted repository which appears to have been since
locked down. To avoid future issues I forked the repository in the
SimpliVity Github account and made it build into our local repo.

Upgraded the wrapper to version 2.10 just to match our own versions.